requirements.txt
Flask== 1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
Jinja2==2.11.1
requests==2.23.0
SQLAlchemy==1.3.13
WTForms==2.2.1
gunicorn==20.0.4
Flask-Script==2.0.6
flask-login==0.5.0
flask-cors==3.0.8
paho-mqtt==1.5.0